face-id-login/backend/requirements.txt
flask
flask-cors
face_recognition
pillow
numpy
mysql-connector-python
